Split lib.rs into modules
